code: move overlays into their own file
